Imported Upstream version 5.1.0~BETA3-dfsg

+++ src/VBox/Devices/Graphics/DevVGA_VBVA.cpp
@@ -2415,7 +2415,7 @@ static int vbvaHandleQueryModeHints(PVGASTATE pVGAState, const VBVAQUERYMODEHINT
         memcpy(pbHint, &pCtx->aModeHints[iHint],
                RT_MIN(parms.cbHintStructureGuest, sizeof(VBVAMODEHINT)));
         pbHint += parms.cbHintStructureGuest;
-        Assert(pbHint - (uint8_t *)pQueryModeHints <= cbBuffer);
+        Assert((uintptr_t)(pbHint - (uint8_t *)pQueryModeHints) <= cbBuffer);
     }
 
     return VINF_SUCCESS;
